third_party/WebKit/LayoutTests/fast/loader/invalid-module-specifier.html

<script>
// This test is Blink specific since it checks error message text.
if (window.testRunner)
  testRunner.dumpAsText();
</script>
<script type="module">
import "http://example.com:invalid-port-num";
</script>
<script type="module">
import "specifier-without-valid-prefix";
</script>
