third_party/WebKit/LayoutTests/fast/loader/invalid-module-specifier-expected.txt

CONSOLE ERROR: Uncaught TypeError: Failed to resolve module specifier "http://example.com:invalid-port-num".
CONSOLE ERROR: Uncaught TypeError: Failed to resolve module specifier "specifier-without-valid-prefix". Note that relative references must start with either "/", "./", or "../".
